Updated Step 2 of adding a safe with module address input of Step 3
