+++ src/modules/dashboard/components/Dialogs/AddExistingSafeDialog/AddExistingSafeDialogForm.css
@@ -48,14 +48,24 @@
 }
 
 .instructions {
+  padding-bottom: 18px;
+  border-bottom: 1px solid color-mod(var(--temp-grey-blue-7) alpha(15%));
   font-size: var(--size-normal);
   font-weight: var(--weight-normal);
   color: var(--temp-grey-blue-7);
 }
 
+.step3Instructions {
+  margin-bottom: -20px;
+  padding-bottom: 0;
+  border: none;
+}
+
 .moduleLinkSection {
   display: flex;
   flex-direction: column;
+  padding-bottom: 0;
+  border: none;
 }
 
 .copyable {
@@ -95,9 +105,11 @@
 .warning {
   display: flex;
   justify-content: center;
-  padding: 15px 0;
+  padding: 14px;
   border-radius: 5px;
-  background-color: color-mod(var(--pink) alpha(15%));
+  background-color: color-mod(var(--golden) alpha(15%));
+  line-height: 18px;
+  color: var(--dark);
   column-gap: 5px;
 }
 
@@ -158,4 +170,27 @@
   position: absolute;
   right: 10px;
   bottom: 28px;
+}
+
+.learnMoreLink {
+  margin-right: 4px;
+  position: relative;
+  top: 1px;
+}
+
+.moduleContractAddressContainer {
+  margin-top: 10px;
+  padding-top: 20px;
+  border-top: 1px solid color-mod(var(--temp-grey-blue-7) alpha(15%));
+}
+
+.warningIcon {
+  margin-right: 5px;
+  height: 18px;
+  width: 18px;
+  fill: var(--golden);
+}
+
+.moduleAddressSubtitle {
+  margin-bottom: 10px;
 }
