+++ src/modules/dashboard/components/Dialogs/AddExistingSafeDialog/AddExistingSafeDialogForm.css.d.ts
@@ -5,6 +5,7 @@ export const step1Subtitle: string;
 export const chainSelect: string;
 export const headingContainer: string;
 export const instructions: string;
+export const step3Instructions: string;
 export const moduleLinkSection: string;
 export const copyable: string;
 export const fat: string;
@@ -17,3 +18,7 @@ export const chainName: string;
 export const safe: string;
 export const moduleLabel: string;
 export const avatarIcon: string;
+export const learnMoreLink: string;
+export const moduleContractAddressContainer: string;
+export const warningIcon: string;
+export const moduleAddressSubtitle: string;
